pridanie oznamu s fotkou, spravne vymazanie podla uzivatelov, zobrazenie oznamu bez fotky

+++ app/src/main/java/com/example/projectsoc/NoteAdapter.java
@@ -1,20 +1,29 @@
 package com.example.projectsoc;
 
+import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.firebase.ui.firestore.FirestoreRecyclerAdapter;
 import com.firebase.ui.firestore.FirestoreRecyclerOptions;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentSnapshot;
+import com.squareup.picasso.Picasso;
+
+import es.dmoral.toasty.Toasty;
 
 public class NoteAdapter extends FirestoreRecyclerAdapter<Note, NoteAdapter.NoteHolder> {
 
+    private View v;
     private OnItemClickListener listener;
+    private FirebaseAuth mAuth = FirebaseAuth.getInstance();
 
     public NoteAdapter(@NonNull FirestoreRecyclerOptions<Note> options) {
         super(options);
@@ -24,32 +33,41 @@ public class NoteAdapter extends FirestoreRecyclerAdapter<Note, NoteAdapter.Note
     protected void onBindViewHolder(@NonNull NoteHolder noteHolder, int i, @NonNull Note note) {
         noteHolder.textViewTitle.setText(note.getTitle());
         noteHolder.textViewDescription.setText(note.getDescription());
-        noteHolder.textViewPhoneNumber.setText(note.getPhoneNumber());
+        if (note.getUrlImage().equals("@drawable/ic_pet")){
+            noteHolder.imageViewDog.setImageResource(R.drawable.ic_pet);
+        }else {
+            Picasso.with(v.getContext()).load(Uri.parse(note.getUrlImage())).fit().centerCrop().into(noteHolder.imageViewDog);
+        }
     }
 
     @NonNull
     @Override
     public NoteHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View v = LayoutInflater.from(parent.getContext()).inflate(R.layout.note_item,parent,false);
+        v = LayoutInflater.from(parent.getContext()).inflate(R.layout.note_item,parent,false);
         return new NoteHolder(v);
     }
 
-    public  void deleteItem(int position){
-        getSnapshots().getSnapshot(position).getReference().delete();
-
+    public void deleteItem(int position){
+        Note note = getSnapshots().getSnapshot(position).toObject(Note.class);
+        String uID = mAuth.getCurrentUser().getUid();
+        if(uID.equals(note.getUserID())) {
+            getSnapshots().getSnapshot(position).getReference().delete();
+        }else{
+            Toasty.error(v.getContext(),"Tento oznam si nepridal ty", Toast.LENGTH_SHORT).show();
+        }
     }
 
     class NoteHolder extends RecyclerView.ViewHolder {
 
         TextView textViewTitle;
         TextView textViewDescription;
-        TextView textViewPhoneNumber;
+        ImageView imageViewDog;
 
         public NoteHolder(@NonNull View itemView) {
             super(itemView);
             textViewTitle = itemView.findViewById(R.id.title);
             textViewDescription = itemView.findViewById(R.id.description);
-            textViewPhoneNumber = itemView.findViewById(R.id.phone_number);
+            imageViewDog = itemView.findViewById(R.id.image);
 
             itemView.setOnClickListener(new View.OnClickListener() {
                 @Override
